+++ mermithid/misc/__init__.py
@@ -8,6 +8,5 @@ from . import KuriePlotTools
 from . import TritiumFormFactor
 from . import FakeTritiumDataFunctions
 from . import ConversionFunctions
-from . import SensitivityFormulas
-from . import SensitivityCavityFormulas
-from . import HannekeFunctions
+from . import CRESFunctions_numericalunits
+from . import Constants_numericalunits
