+++ mermithid/processors/Sensitivity/SensitivityCurveProcessor.py
@@ -27,7 +27,7 @@ ppb = 1e-9
 # morpho imports
 from morpho.utilities import morphologging, reader
 from morpho.processors import BaseProcessor
-from mermithid.misc.SensitivityFormulas import Sensitivity
+from mermithid.sensitivity.SensitivityFormulas import Sensitivity
 
 
 logger = morphologging.getLogger(__name__)
