+++ mermithid/sensitivity/SensitivityCavityFormulas.py
@@ -9,48 +9,11 @@ CDR (CRES design report, Section 1.3) https://www.overleaf.com/project/5b9314afc
 import numpy as np
 import configparser
 from numpy import pi
-from mermithid.misc.HannekeFunctions import *
 
-# Numericalunits is a package to handle units and some natural constants
-# natural constants
-from numericalunits import e, me, c0, eps0, kB, hbar
-from numericalunits import meV, eV, keV, MeV, cm, m, ns, s, Hz, kHz, MHz, GHz, amu, nJ
-from numericalunits import nT, uT, mT, T, mK, K,  C, F, g, W
-from numericalunits import hour, year, day, s, ms
-from numericalunits import mu0, NA, kB, hbar, me, c0, e, eps0, hPlanck
+from mermithid.misc.Constants_numericalunits import *
+from mermithid.misc.CRESFunctions_numericalunits import *
+from mermithid.cavity.HannekeFunctions import *
 
-T0 = -273.15*K
-
-tritium_livetime = 5.605e8*s
-tritium_mass_atomic = 3.016* amu *c0**2
-tritium_electron_crosssection_atomic = 9.e-23*m**2 #Hamish extrapolated to 18.6keV using Shah et al. (1987): https://iopscience.iop.org/article/10.1088/0022-3700/20/14/022
-tritium_endpoint_atomic = 18563.251*eV
-last_1ev_fraction_atomic = 2.067914e-13/eV**3
-
-tritium_mass_molecular = 6.032099 * amu *c0**2
-tritium_electron_crosssection_molecular = 3.67*1e-22*m**2 #[Inelastic from Aseev (2000) for T2] + [Elastic from Liu (1987) for H2, extrapolated by Elise to 18.6keV]
-tritium_endpoint_molecular = 18574.01*eV
-last_1ev_fraction_molecular = 1.67364e-13/eV**3
-
-ground_state_width = 0.436 * eV
-#ground_state_width_uncertainty = 0.001*0.436*eV
-
-gyro_mag_ratio_proton = 42.577*MHz/T
-
-# units that do not show up in numericalunits
-# missing pre-factors
-fW = W*1e-15
-
-# unitless units, relative fractions
-pc = 0.01
-ppm = 1e-6
-ppb = 1e-9
-ppt = 1e-12
-ppq = 1e-15
-
-# radian and degree which are also not really units
-rad = 1
-deg = np.pi/180
 
 
 try:
@@ -70,41 +33,7 @@ class NameSpace(object):
 
 ##############################################################################
 # CRES functions
-def gamma(kin_energy):
-    return kin_energy/(me*c0**2) + 1
 
-def beta(kin_energy):
-    # electron speed at kin_energy
-    return np.sqrt(kin_energy**2+2*kin_energy*me*c0**2)/(kin_energy+me*c0**2)
-
-def frequency(kin_energy, magnetic_field):
-    # cyclotron frequency
-    return e/(2*np.pi*me)/gamma(kin_energy)*magnetic_field
-
-def wavelength(kin_energy, magnetic_field):
-    return c0/frequency(kin_energy, magnetic_field)
-
-def kin_energy(freq, magnetic_field):
-    return (e*c0**2/(2*np.pi*freq)*magnetic_field - me*c0**2)
-
-def rad_power(kin_energy, pitch, magnetic_field):
-    # electron radiation power
-    f = frequency(kin_energy, magnetic_field)
-    b = beta(kin_energy)
-    Pe = 2*np.pi*(e*f*b*np.sin(pitch/rad))**2/(3*eps0*c0*(1-b**2))
-    return Pe
-
-def track_length(rho, kin_energy=None, molecular=True):
-    if kin_energy is None:
-        kin_energy = tritium_endpoint_molecular if molecular else tritium_endpoint_atomic
-    crosssect = tritium_electron_crosssection_molecular if molecular else tritium_electron_crosssection_atomic
-    return 1 / (rho * crosssect * beta(kin_energy) * c0)
-
-def sin2theta_sq_to_Ue4_sq(sin2theta_sq):
-    return 0.5*(1-np.sqrt(1-sin2theta_sq**2))
-
-def Ue4_sq_to_sin2theta_sq(Ue4_sq):
-    return 4*Ue4_sq*(1-Ue4_sq)
 
 # Wouters functinos
 def db_to_pwr_ratio(q_db):
